+++ src/main/java/com/kh/vira_dev/ecommerceapi/dto/response/AddressResponse.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Builder
 public class AddressResponse {
 
-    private Short addressId;
+    private Long addressId;
 
     private String fullName;
